pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "llm-router"
version = "0.1.0"
description = "LLM API Router with OpenAI protocol support and MCP XML tool parsing"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "ansatz Vibe"},
]
keywords = ["flask", "llm", "openai", "mcp", "tool-calling", "router"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Internet :: WWW/HTTP :: WSGI :: Application",
]
dependencies = [
    "flask>=3.0.0",
    "json-repair>=0.30.0",
    "python-dotenv>=1.0.0",
    "openai>=1.0.0",
    "regex>=2024.0.0",
    "gunicorn>=21.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
llm-router = "llm_router.cli:main"

[project.urls]
Homepage = "https://github.com/yourusername/llm-router"
Repository = "https://github.com/yourusername/llm-router"
Issues = "https://github.com/yourusername/llm-router/issues"

[tool.setuptools.packages.find]
include = ["llm_router*"]
exclude = ["tests*", "examples*"]

[tool.setuptools.package-data]
llm_router = ["py.typed"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]
ignore = ["E501"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.coverage.run]
source = ["llm_router"]
